Change yocto uid and gid
This change is the counter part for the commit be96024a9da54b35058aa6fef72a94e665a8441d
in the yocto-image repository. This change applies the new uid and gid
and removes the unnecessary usermod instructions from the readme.
commit be96024a9da54b35058aa6fef72a94e665a8441d:
change the custom uid and gid to the standard uid/gid for the first
user on a fedora image. This helps to reduce changes done to the upstream
distro image.

Signed-off-by: Lars Niesen <lars.niesen@gmx.de>